chore: modify invoice API reponse

+++ backend/src/main/java/com/dashboard/backend/dto/PendingAmountDTO.java
@@ -1,3 +1,6 @@
 package com.dashboard.backend.dto;
 
-public record PendingAmountDTO(long pendingAmount) {}
+import java.math.BigDecimal;
+
+public record PendingAmountDTO(BigDecimal pendingAmount) {
+}
